Skip resolved classnames that contain a dot
For some reason, sometimes structures that are psr-0 compatible contain folders starting with a dot - in my case it was .AppleDouble. Preventing this would have a been a huge pain, so I figured it could be a good idea to do this. If could be enhanced by a whole set of blacklisted characters, or a regex whitelisting the format of allowed classnames in php.