Add --quiet option to suppress STDOUT output
we want to suppress any output that is not an error when running
pdepend. For this I am adding a new option --quiet to the TextUI
Command. Setting this flag will put pdepend into silent mode, meaning it
will use the NulLResultPrinter instead of the regular ResultPrinter and
it will not output version, workarounds and runtime statistics. It will
output any errors though.

Closes #211