(fix): match expression does not have namespace
the keyword `match` is being treated as if it could have a namespace,
but in PHP 8 you can't use `match` in your code, given it is a reserved
word now.

this fix does keep match as a Tokens::T_STRING, but if the maintainers
prefer that I change it to use a new token id (Tokens::T_MATCH), I can
invest more time to change and fix that way.