Declare ReturnTypeWillChange on internal overloads
Proposal to add the PHP Attribute `#[\ReturnTypeWillChange]` onto each method that overwrites internal methods for PHP 8.1.

As of PHP 8.1, the return types of these methods must be defined to not be understood as `: mixed`, and be incompatible with internal definition. Due to the fact pDepend must continue to work for PHP >= 5.3.7, we can't add the return type definitions as this is a feature for PHP >= 7. That's why I propose the use of this attribute.

Due to the format of PHP Attributes syntax, this edition will have no effect on PHP < 8. On PHP 8.0, the attribute `ReturnTypeWillChange` has no effect. This is only to be compatible with PHP >= 8.1 and PHP < 9.